KAT-3271 Record videos are not generated in custom report's location when executing a Test Suite in console mode execution

+++ source/com.kms.katalon.execution/src/com/kms/katalon/execution/launcher/ReportableLauncher.java
@@ -219,8 +219,14 @@ public abstract class ReportableLauncher extends LoggableLauncher {
                     }
 
                     // Copy child file to user's report folder
-                    FileUtils.copyFile(reportChildSourceFile,
-                            new File(userReportFolder, fileName + "." + fileExtension));
+                    if (reportChildSourceFile.isFile()) {
+                        FileUtils.copyFile(reportChildSourceFile,
+                                new File(userReportFolder, fileName + "." + fileExtension));
+                    } else if (reportChildSourceFile.isDirectory()) {
+                        File newCoppiedFolder = new File(userReportFolder, fileName);
+                        newCoppiedFolder.mkdirs();
+                        FileUtils.copyDirectory(reportChildSourceFile, newCoppiedFolder);
+                    }
                 }
             }
         } catch (IOException ex) {
